All python files and the dataset
The dataset is in the data_with_categories.rar file. the mp_cat_before_pickle.py is used to train the program until an OK accuracy is aquired. As I am pickling everything in the program, once the accuracy is obtained, if the pickled classifier is loaded, it will posses the same accuracy where we stopped training. mp_cat.py is used to categorize the text sent to it into one of the six categories. subjects.py is used to access the emails from gmail and send the body to mp_cat.py for categorization. It also creates a local database to store all the attributes of emails like subject, sender etc.